requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==12.1.0
django-bootstrap4==1.1.1
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==3.0.2
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
numpy==1.15.2
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
pyuploadcare==2.7.0
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
whitenoise==5.1.0




